+++ src/features/translation/providers/GoogleTranslateV2Provider.js
@@ -3,6 +3,7 @@ import { getScopedLogger } from '@/shared/logging/logger.js';
 import { LOG_COMPONENTS } from '@/shared/logging/logConstants.js';
 import { ProviderNames } from "@/features/translation/providers/ProviderConstants.js";
 import { TraditionalTextProcessor, getTextInfo } from "./utils/TraditionalTextProcessor.js";
+import { isolateParagraphChunks } from "./utils/ParagraphChunkIsolation.js";
 import { TRANSLATION_CONSTANTS } from "@/shared/config/translationConstants.js";
 import {
   getProviderLanguageCode
@@ -20,7 +21,6 @@ import {
 import { getTranslationString } from "@/utils/i18n/i18n.js";
 
 const logger = getScopedLogger(LOG_COMPONENTS.PROVIDERS, 'GoogleTranslateV2');
-const PARAGRAPH_BREAK_REGEX = /\n{2,}/;
 
 /**
  * Stable TKK value for Google Translate token generation.
@@ -54,55 +54,7 @@ export class GoogleTranslateV2Provider extends BaseTranslateProvider {
 
   async _createChunks(texts) {
     const chunks = await super._createChunks(texts);
-    return this._isolateParagraphChunks(chunks);
-  }
-
-  _isolateParagraphChunks(chunks) {
-    const isolatedChunks = [];
-
-    for (const chunk of chunks) {
-      if (!chunk?.texts?.some(item => PARAGRAPH_BREAK_REGEX.test(getTextInfo(item).text))) {
-        isolatedChunks.push(chunk);
-        continue;
-      }
-
-      let currentTexts = [];
-
-      const flushCurrent = () => {
-        if (currentTexts.length === 0) return;
-        isolatedChunks.push({
-          ...chunk,
-          texts: currentTexts,
-          charCount: this._calculateChunkCharCount(currentTexts)
-        });
-        currentTexts = [];
-      };
-
-      for (const item of chunk.texts) {
-        const info = getTextInfo(item);
-        if (PARAGRAPH_BREAK_REGEX.test(info.text)) {
-          flushCurrent();
-          isolatedChunks.push({
-            ...chunk,
-            texts: [item],
-            charCount: info.length
-          });
-          continue;
-        }
-
-        currentTexts.push(item);
-      }
-
-      flushCurrent();
-    }
-
-    return isolatedChunks;
-  }
-
-  _calculateChunkCharCount(texts) {
-    const delimiterLength = TRANSLATION_CONSTANTS.TEXT_DELIMITER?.length || 0;
-    const textLength = texts.reduce((sum, item) => sum + getTextInfo(item).length, 0);
-    return textLength + (Math.max(0, texts.length - 1) * delimiterLength);
+    return isolateParagraphChunks(chunks);
   }
 
   _getLangCode(lang) {
